# specs/001-invoice-management/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Invoice Management API
  version: 1.0.0
  description: REST contract for the first-release invoice management application.
servers:
  - url: /api/v1
paths:
  /customers:
    post:
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CustomerCreate' }
      responses:
        '201': { description: Created customer, content: { application/json: { schema: { $ref: '#/components/schemas/Customer' } } } }
        '422': { $ref: '#/components/responses/ValidationError' }
    get:
      summary: List customers
      responses:
        '200': { description: Customer list, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Customer' } } } } }
  /customers/{customerId}:
    get:
      summary: Get a customer
      parameters: [ { $ref: '#/components/parameters/CustomerId' } ]
      responses:
        '200': { description: Customer, content: { application/json: { schema: { $ref: '#/components/schemas/Customer' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /invoices:
    post:
      summary: Create a Draft invoice and derive all totals
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InvoiceInput' }
      responses:
        '201': { description: Created invoice, content: { application/json: { schema: { $ref: '#/components/schemas/Invoice' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    get:
      summary: List invoices
      responses:
        '200': { description: Invoice list, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/InvoiceListItem' } } } } }
  /invoices/{invoiceId}:
    get:
      summary: Get an invoice
      parameters: [ { $ref: '#/components/parameters/InvoiceId' } ]
      responses:
        '200': { description: Invoice detail, content: { application/json: { schema: { $ref: '#/components/schemas/Invoice' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Replace editable fields and items on a Draft invoice
      parameters: [ { $ref: '#/components/parameters/InvoiceId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InvoiceInput' }
      responses:
        '200': { description: Updated invoice, content: { application/json: { schema: { $ref: '#/components/schemas/Invoice' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /invoices/{invoiceId}/status:
    post:
      summary: Apply an allowed invoice status transition
      parameters: [ { $ref: '#/components/parameters/InvoiceId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties: { status: { $ref: '#/components/schemas/InvoiceStatus' } }
      responses:
        '200': { description: Transitioned invoice, content: { application/json: { schema: { $ref: '#/components/schemas/Invoice' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
components:
  parameters:
    CustomerId: { name: customerId, in: path, required: true, schema: { type: integer, minimum: 1 } }
    InvoiceId: { name: invoiceId, in: path, required: true, schema: { type: integer, minimum: 1 } }
  responses:
    NotFound: { description: Resource not found }
    Conflict: { description: Invalid state; response contains a code and message }
    ValidationError: { description: Invalid request data }
  schemas:
    CustomerCreate:
      type: object
      required: [name, phone, email, address]
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string, minLength: 1 }
        email: { type: string, format: email }
        address: { type: string, minLength: 1 }
    Customer:
      allOf:
        - $ref: '#/components/schemas/CustomerCreate'
        - type: object
          required: [id]
          properties: { id: { type: integer } }
    InvoiceItemInput:
      type: object
      required: [description, quantity, unit_price]
      properties:
        description: { type: string, minLength: 1 }
        quantity: { type: string, pattern: '^\\d+(\\.\\d{1,3})?$' }
        unit_price: { type: string, pattern: '^\\d+(\\.\\d{1,2})?$' }
    InvoiceInput:
      type: object
      required: [customer_id, issue_date, due_date, items]
      properties:
        customer_id: { type: integer, minimum: 1 }
        issue_date: { type: string, format: date }
        due_date: { type: string, format: date }
        items: { type: array, minItems: 1, items: { $ref: '#/components/schemas/InvoiceItemInput' } }
        notes: { type: string }
        payment_terms: { type: string }
    InvoiceStatus:
      type: string
      enum: [Draft, Sent, Paid, Overdue, Cancelled]
    InvoiceItem:
      allOf:
        - $ref: '#/components/schemas/InvoiceItemInput'
        - type: object
          required: [line_total]
          properties: { line_total: { type: string, pattern: '^\\d+\\.\\d{2}$' } }
    InvoiceListItem:
      type: object
      required: [id, invoice_number, customer, issue_date, due_date, status, total]
      properties:
        id: { type: integer }
        invoice_number: { type: string, pattern: '^INV-\\d{4,}$' }
        customer: { $ref: '#/components/schemas/Customer' }
        issue_date: { type: string, format: date }
        due_date: { type: string, format: date }
        status: { $ref: '#/components/schemas/InvoiceStatus' }
        total: { type: string, pattern: '^\\d+\\.\\d{2}$' }
    Invoice:
      allOf:
        - $ref: '#/components/schemas/InvoiceListItem'
        - type: object
          required: [items, subtotal, tax, notes, payment_terms]
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/InvoiceItem' } }
            subtotal: { type: string, pattern: '^\\d+\\.\\d{2}$' }
            tax: { type: string, pattern: '^\\d+\\.\\d{2}$' }
            notes: { type: string }
            payment_terms: { type: string }